# docs/openapi/rgs_casino.yaml
openapi: 3.0.0
info:
  title: RGS⇄Casino Interface API
  version: 1.0.0
  description: Complete RGS and Casino communication interface (21 endpoints)
servers:
  - url: http://localhost:3001/casino/api/v1
    description: Casino API
  - url: http://localhost:3001/rgs/api/v1
    description: RGS API
paths:
  /wallet/debit:
    post:
      summary: Debit player wallet
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId, amount, currency]
              properties:
                playerId:
                  type: string
                  format: uuid
                amount:
                  type: string
                  pattern: '^\d+\.\d{2}$'
                currency:
                  type: string
                  enum: [EUR]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /wallet/credit:
    post:
      summary: Credit player wallet
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/WalletOperation'
      responses:
        '200':
          $ref: '#/components/responses/WalletSuccess'
        '422':
          $ref: '#/components/responses/ValidationError'
  /wallet/cancel:
    post:
      summary: Cancel wallet transaction
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId, transactionId, amount]
              properties:
                playerId:
                  type: string
                  format: uuid
                transactionId:
                  type: string
                  format: uuid
                amount:
                  type: string
                  pattern: '^\d+\.\d{2}$'
                reason:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/WalletSuccess'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /player/status:
    get:
      summary: Get player status (canonical)
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: playerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Player status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerStatus'
    post:
      summary: Get player status (deprecated)
      deprecated: true
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId]
              properties:
                playerId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Player status
          headers:
            Deprecation:
              schema:
                type: string
                enum: ['true']
            Sunset:
              schema:
                type: string
                format: date-time
            Link:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerStatus'
  /kyc/status:
    get:
      summary: Get KYC status (canonical)
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: playerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: KYC status
    post:
      summary: Get KYC status (deprecated)
      deprecated: true
      tags: [Casino]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: KYC status with deprecation headers
  /kyc/exclusions:
    get:
      summary: Get exclusions (canonical)
      tags: [Casino]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: playerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exclusions list
    post:
      summary: Get exclusions (deprecated)
      deprecated: true
      tags: [Casino]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Exclusions with deprecation headers
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: UUID v4 for request tracing
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: UUID v4 for idempotent operations (48h window)
  schemas:
    WalletResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED]
        reason:
          type: string
          pattern: '^RW-'
        newBalance:
          type: string
          pattern: '^\d+\.\d{2}$'
    PlayerStatus:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        status:
          type: string
        kyc:
          type: object
        rgFlags:
          type: object
        activeSessions:
          type: array
  requestBodies:
    WalletOperation:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [playerId, amount, currency]
            properties:
              playerId:
                type: string
                format: uuid
              amount:
                type: string
                pattern: '^\d+\.\d{2}$'
              currency:
                type: string
                enum: [EUR]
...
  responses:
    WalletSuccess:
      description: Wallet operation successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WalletResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code:
                    type: string
                    pattern: '^RW-'
                  message:
                    type: string
              correlationId:
                type: string
                format: uuid